feat: add .golangci.yml

+++ scripts/tasks/Show-Clock.ps1
@@ -31,5 +31,9 @@ else {
 Register-PSScheduledTask 'Stop-Computer' {
   Send-Notify -Title 'Stop-Computer' -Severity Error 'Computer will be force stop after 3 minutes.'
   Start-Sleep 0:3
-  Stop-Computer -Force
+  Stop-Computer
+  if ($IsWindows) {
+    Start-Sleep 0:3
+    Stop-Computer -Force
+  }
 } -DaysInterval 1 -At 23:27 -UsePowerShell -Force
